complete

+++ Math/Program.cs
@@ -6,12 +6,12 @@ class Operation
 {
     public static double Add(double x, double y)
     { return x + y; }
-    public static int Subtract(int x, int y)
+    public static double Subtract(double x, double y)
     { return x - y; }
 
-    public static int Multiply(int x, int y)
+    public static double Multiply(double x, double y)
     { return x * y; }
 
-    public static double Divide(int x, int y)
+    public static double Divide(double x, double y)
     { return (double)x / y; }
 }
